Jinx/data_service/requirements.txt
fastapi
celery
uvicorn
pydantic
python-jose
python-dotenv
